pyproject.toml
[tool.poetry]
name = "django-getpaid-payu"
version = "0.2.0-pre.3"
description = "Django-GetPaid plugin for PayU service."
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/django-getpaid/django-getpaid-payu"
packages = [
    { include = "getpaid_payu" }
]


[tool.poetry.dependencies]
python = "^3.6"
django-getpaid = "^2.1.0-pre.1"
requests = "^2.23.0"
swapper = "^1.1.2"
typing-extensions = "^3.7.4"


[tool.poetry.dev-dependencies]
pre-commit = "^2.1.1"
factory-boy = "^2.12.0"
black = "^19.10b0"
tox = "^3.14.5"
ipdb = "^0.13.2"
pytest = {version = "^5.3.5", optional = true}
codecov = {version = "^2.0.16", optional = true}
coverage = {version = "^5.0.3", optional = true}
requests-mock = {version = "^1.7.0", optional = true}
pytest-cov = {version = "^2.8.1", optional = true}
pytest-django = {version = "^3.8.0", optional = true}


[tool.poetry.extras]
test = ["pytest", "codecov", "coverage", "request-mock", "pytest-cov", "pytest-django"]


[tool.black]
line-length = 88
target-version = ['py36']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.cache
    | \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | __pycache__
    | _build
    | buck-out
    | build
    | dist
    | docs
    | migrations
  )/

)
'''


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 88
known_first_party = ["getpaid_payu"]
known_third_party = ["django", "django_fsm", "factory", "getpaid", "orders", "paywall", "pendulum", "pytest", "pytest_factoryboy", "requests", "swapper", "typing_extensions"]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
